Fixed #199
Changed login window in login.css to no longer get "squished" on a smaller window by changing the margin-right in .login-form-wrapper to be the relative 45wh instead of the absolute 1111px.

Changed signup window in signup.css to no longer get *as* "squished" on a smaller window by changing the gap in .signup-form.form-group and margin-right in .signup-container to relative values.